Ajout de 'dora.oidc` en tant qu'app Django
Pour modularisation.
Voir commentaires dans `__init__.py`.
